revert structural changes
